requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
accelerate==1.6.0
    # via transformers
aiofiles==24.1.0
    # via gradio
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via litellm
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   gradio
    #   httpx
    #   openai
    #   starlette
asyncio==3.4.3
    # via learn-general-ai-agent (pyproject.toml)
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
authlib==1.5.2
    # via gradio
beautifulsoup4==4.13.4
    # via markdownify
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   soundfile
charset-normalizer==3.4.1
    # via requests
chess==1.11.2
    # via learn-general-ai-agent (pyproject.toml)
click==8.1.8
    # via
    #   duckduckgo-search
    #   litellm
    #   typer
    #   uvicorn
colorama==0.4.6
    # via py-mon
cryptography==44.0.2
    # via authlib
distro==1.9.0
    # via openai
dotenv==0.9.9
    # via learn-general-ai-agent (pyproject.toml)
duckduckgo-search==8.0.1
    # via smolagents
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.115.12
    # via gradio
ffmpeg==1.4
    # via learn-general-ai-agent (pyproject.toml)
ffmpy==0.5.0
    # via gradio
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.2
    # via
    #   gradio-client
    #   huggingface-hub
    #   torch
gradio==5.27.0
    # via learn-general-ai-agent (pyproject.toml)
gradio-client==1.9.0
    # via gradio
groovy==0.1.2
    # via gradio
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   gradio
    #   gradio-client
    #   litellm
    #   openai
    #   safehttpx
huggingface-hub==0.30.2
    # via
    #   accelerate
    #   gradio
    #   gradio-client
    #   smolagents
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
itsdangerous==2.2.0
    # via gradio
jinja2==3.1.6
    # via
    #   gradio
    #   litellm
    #   smolagents
    #   torch
jiter==0.9.0
    # via openai
jsonschema==4.23.0
    # via litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.67.4.post1
    # via
    #   learn-general-ai-agent (pyproject.toml)
    #   smolagents
lxml==5.4.0
    # via duckduckgo-search
markdown-it-py==3.0.0
    # via rich
markdownify==1.1.0
    # via smolagents
markupsafe==3.0.2
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numpy==2.2.5
    # via
    #   accelerate
    #   gradio
    #   pandas
    #   smolagents
    #   soundfile
    #   torchvision
    #   transformers
openai==1.76.0
    # via litellm
openpyxl==3.1.5
    # via learn-general-ai-agent (pyproject.toml)
orjson==3.10.16
    # via gradio
packaging==25.0
    # via
    #   accelerate
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   transformers
panda==0.3.1
    # via learn-general-ai-agent (pyproject.toml)
pandas==2.2.3
    # via gradio
pathlib==1.0.1
    # via learn-general-ai-agent (pyproject.toml)
pillow==11.2.1
    # via
    #   gradio
    #   smolagents
    #   torchvision
primp==0.15.0
    # via duckduckgo-search
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via accelerate
py-mon==2.1.0
    # via learn-general-ai-agent (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   fastapi
    #   gradio
    #   litellm
    #   openai
pydantic-core==2.33.1
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   dotenv
    #   litellm
    #   smolagents
python-multipart==0.0.20
    # via gradio
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   gradio
    #   huggingface-hub
    #   transformers
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests==2.32.3
    # via
    #   learn-general-ai-agent (pyproject.toml)
    #   huggingface-hub
    #   panda
    #   smolagents
    #   tiktoken
    #   transformers
rich==14.0.0
    # via
    #   smolagents
    #   typer
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
ruff==0.11.7
    # via
    #   learn-general-ai-agent (pyproject.toml)
    #   gradio
safehttpx==0.1.6
    # via gradio
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
semantic-version==2.10.0
    # via gradio
setuptools==79.0.1
    # via panda
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
smolagents==1.14.0
    # via learn-general-ai-agent (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soundfile==0.13.1
    # via smolagents
soupsieve==2.7
    # via beautifulsoup4
starlette==0.46.2
    # via
    #   fastapi
    #   gradio
sympy==1.14.0
    # via torch
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.1
    # via
    #   litellm
    #   transformers
tomlkit==0.13.2
    # via gradio
torch==2.7.0
    # via
    #   accelerate
    #   smolagents
    #   torchvision
    #   transformers
torchvision==0.22.0
    # via smolagents
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==4.51.3
    # via learn-general-ai-agent (pyproject.toml)
typer==0.15.2
    # via gradio
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via gradio
watchdog==6.0.0
    # via py-mon
websockets==15.0.1
    # via gradio-client
yarl==1.20.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
